SAKY-201 : Added structs and constructor function

+++ main.go
@@ -5,13 +5,13 @@ import "fmt"
 //type taxPricesMap
 
 func main() {
-	prices := []float32{10, 20, 30}
-	taxRates := []float32{0, 0.07, 0.1, 0.15}
+	prices := []float64{10, 20, 30}
+	taxRates := []float64{0, 0.07, 0.1, 0.15}
 
-	taxRatePricesMap := make(map[float32][]float32)
+	taxRatePricesMap := make(map[float64][]float64)
 
 	for _, taxRate := range taxRates {
-		taxPrice := make([]float32, len(prices))
+		taxPrice := make([]float64, len(prices))
 		for j, price := range prices {
 			taxPrice[j] = price * (1 + taxRate)
 		}
